requirements.txt
Flask==3.0.0
Flask-MySQLdb==2.0.1
Flask-Mail==0.9.1
itsdangerous==2.1.2
web3==6.11.1
py-solc-x==2.0.3
